// apps/web-user/src/config/koudamaIconAssignments.generated.ts
export type KoudamaIconName =
  | 'BadgeDollarSign'
  | 'BadgePlus'
  | 'Bell'
  | 'Bookmark'
  | 'BotMessageSquare'
  | 'BriefcaseBusiness'
  | 'Calculator'
  | 'CalendarDays'
  | 'Car'
  | 'CedarTreeFull'
  | 'ChartNoAxesColumn'
  | 'CircleDollarSign'
  | 'CircleHelp'
  | 'Clapperboard'
  | 'ClipboardList'
  | 'Compass'
  | 'FilePenLine'
  | 'FileText'
  | 'Flag'
  | 'FolderOpen'
  | 'Gavel'
  | 'GraduationCap'
  | 'HandCoins'
  | 'Handshake'
  | 'HandsHeart'
  | 'Home'
  | 'Inbox'
  | 'Landmark'
  | 'Layers'
  | 'LayoutGrid'
  | 'LifeBuoy'
  | 'Link'
  | 'List'
  | 'LogIn'
  | 'MailOpen'
  | 'MapPinned'
  | 'Megaphone'
  | 'MessageCircle'
  | 'MessagesSquare'
  | 'Mic'
  | 'MoreHorizontalCircle'
  | 'Newspaper'
  | 'Pencil'
  | 'PlusCircle'
  | 'ReceiptText'
  | 'Search'
  | 'Settings'
  | 'ShieldCheck'
  | 'ShoppingCart'
  | 'Siren'
  | 'Sparkles'
  | 'Store'
  | 'Ticket'
  | 'TriangleAlert'
  | 'Trophy'
  | 'UserCheck'
  | 'UserRound'
  | 'UsersRound'
  | 'Vote'
  | 'Wrench'